+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 # 1.1.5 - In progress
 
-- [Feature] Compose Theme
+- [Feature] Compose Theme(new module `core:ui:theme`)
 - [BUGFIX] Install portrait orientation
+
 
 # 1.1.4